Add files via upload
This repository contains an in-depth analysis of multivariate statistical methods applied to a dataset of electric vehicle attributes. The project aims to explore the relationships between multiple variables and their influence on electric vehicle range and popularity. Key statistical techniques, such as multivariable linear regression and ANOVA, are implemented to extract meaningful insights.