Add files via upload
Election2024-project
The Election2024 project is dedicated to analyzing and visualizing data related to the 2024 election using Python for data processing and Power Bi for interactive visualizations. This project aims to serve as a comprehensive tool for researchers, analysts, and the general public to explore voter demographics, polling data, and election outcomes.